Remove references to marathon from check.py; move get_all_namespaces(for_service) from marathon_tools to long_running_service_tools
