+++ paasta_tools/cli/cmds/info.py
@@ -20,8 +20,8 @@ from service_configuration_lib import read_service_configuration
 from paasta_tools.cli.cmds.status import get_actual_deployments
 from paasta_tools.cli.utils import figure_out_service_name
 from paasta_tools.cli.utils import lazy_choices_completer
-from paasta_tools.marathon_tools import get_all_namespaces_for_service
-from paasta_tools.marathon_tools import load_service_namespace_config
+from paasta_tools.long_running_service_tools import get_all_namespaces_for_service
+from paasta_tools.long_running_service_tools import load_service_namespace_config
 from paasta_tools.monitoring_tools import get_runbook
 from paasta_tools.monitoring_tools import get_team
 from paasta_tools.utils import DEFAULT_SOA_DIR
