+++ paasta_tools/firewall.py
@@ -11,7 +11,7 @@ from contextlib import contextmanager
 
 from paasta_tools import iptables
 from paasta_tools.cli.utils import get_instance_config
-from paasta_tools.marathon_tools import get_all_namespaces_for_service
+from paasta_tools.long_running_service_tools import get_all_namespaces_for_service
 from paasta_tools.marathon_tools import marathon_services_running_here
 from paasta_tools.utils import get_running_mesos_docker_containers
 from paasta_tools.utils import load_system_paasta_config
