+++ paasta_tools/generate_services_file.py
@@ -25,8 +25,8 @@ from datetime import datetime
 import service_configuration_lib
 import yaml
 
-from paasta_tools.marathon_tools import get_all_namespaces
-from paasta_tools.marathon_tools import get_all_namespaces_for_service
+from paasta_tools.long_running_service_tools import get_all_namespaces
+from paasta_tools.long_running_service_tools import get_all_namespaces_for_service
 from paasta_tools.utils import atomic_file_write
 from paasta_tools.utils import compose_job_id
 from paasta_tools.utils import DEFAULT_SOA_DIR
