+++ paasta_tools/synapse_srv_namespaces_fact.py
@@ -25,12 +25,12 @@ This is nice to use as a facter fact for Synapse stuff!
 """
 import sys
 
-from paasta_tools import marathon_tools
+from paasta_tools import long_running_service_tools
 
 
 def main():
     strings = []
-    for full_name, config in marathon_tools.get_all_namespaces():
+    for full_name, config in long_running_service_tools.get_all_namespaces():
         if "proxy_port" in config:
             strings.append("{}:{}".format(full_name, config["proxy_port"]))
     strings = sorted(strings)
